Move FeatureProjectionsImp and ProjectionsFromFeatureModel out of Playground to the appropriate places
